src/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
//streamlit>=1.32.0
pydantic-settings>=2.2.1